fix function sha2 parameter unmatch for upgrade

+++ src/include/catalog/upgrade_sql/rollback_catalog_maindb/rollback-post_catalog_maindb_92_806.sql
@@ -1,3 +1,3 @@
 DROP FUNCTION IF EXISTS pg_catalog.sha(text) CASCADE;
 DROP FUNCTION IF EXISTS pg_catalog.sha1(text) CASCADE;
-DROP FUNCTION IF EXISTS pg_catalog.sha2(text, int) CASCADE;
+DROP FUNCTION IF EXISTS pg_catalog.sha2(text, bigint) CASCADE;
